ADD LOGO FILES FOR USE IN ONE PAGE FLYER LISTING
I'm adding the XCF, JPG and PNG files. One thing to do going forwards
may be to track all image files in XCF format and document how to
generate the formats that are actually used from this. Is it possible to
automate this with a script perhaps? Also, we should consider using WebP
as our standard format for distibution.